remove previous database only if it exists

+++ build.sh
@@ -6,6 +6,6 @@ DATABASE="pmdb.db"
 
 (cd ./db-tool/; go build)
 
-rm "$DATABASE"
+[ -e "$DATABASE" ] && rm "$DATABASE"
 cat tables/*.sql | sqlite3 "$DATABASE"
 ./db-tool/db-tool import --db "$DATABASE" --table personal_ratings --src ratings.csv --delim ','
